+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Version 0.10.0-DEV
 - ![INFO][badge-info] Add performance test.
 - ![INFO][badge-info] Update examples.
 - ![INFO][badge-info] Default name for output in ParticleSystem is `default_output_name`, but it can be accessed though `sys.output`, as before.
+- ![INFO][badge-info] Internal fields of ParticleSystems are not part of `PrivateParticleSystemData` and are accessed through internal getter functions, or through `sys.private.field`. 
 - ![BUGFIX][badge-bugfix] Fix cross-computation (`pairwise!(f, x, sys)`) with `NonPeriodicCell` failing to find pairs when particle coordinates span negative values or large coordinate ranges / do not modify input coordinates in `NonPeriodicCell`. This bug was never released, it was created and fixed in the development version.
 - ![BUGFIX][badge-bugfix] If the number of particles is reduced by updating and becomes smaller than nbatches, there was a crash. Fixed. Also fixed automatic updating of nbatches when first set of positions change (bug never released).
 - ![BUGFIX][badge-bugfix] Fix `nbatches` not being updated for the `:map` phase of `CellListPair` systems when the particle count changes (bug never released).